Filter coupon details - ENT-1345
